checkpoints/checkpoint_20251006_140102_start_working_on_the_logbook_review_process_where_the_lgbooks_flows_back_and_forward_between_the_intern_and_supervisor_until_approved/code/backend/competencies/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import CompetencyViewSet

router = DefaultRouter()
router.register(r'competencies', CompetencyViewSet)

urlpatterns = [
    path('api/', include(router.urls)),
]
