checkpoints/checkpoint_20251006_140102_start_working_on_the_logbook_review_process_where_the_lgbooks_flows_back_and_forward_between_the_intern_and_supervisor_until_approved/code/backend/requirements.txt
Django==5.1.2
djangorestframework==3.16.1
django-cors-headers==4.8.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
djangorestframework-simplejwt==5.3.1

